blockchain/requirements.txt
flask
psycopg2-binary
requests
tqdm

# OpenTelemetry
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-psycopg2
